requirements.txt
fastapi
uvicorn
tensorflow
keras
pillow
numpy
opencv-python
python-multipart